修改accuracy的类型，由int转为floatt，在spamEmail.py文件中calAccuracy中的返回值

+++ src/spam/spamEmail.py
@@ -92,4 +92,4 @@ class spamEmailBayes:
                 rightCount+=1
             else:
                 errorCount+=1
-        return rightCount/(rightCount+errorCount)
+        return (float)(rightCount)/(float)(rightCount+errorCount))
